Version 1 of Musica.py
Initial code checkin of Musica.py. This script extracts the contents of a users discogs collections and breaks down the database into the following fields: 
     - Artist (Normalized into Last name, First name format)
     - Title
     - Year (year the recording was released)
     - Genre (discogs uses a free-form method of genre. Users may use their own genre descriptions, usually Jazz, Rock, Classical/Symphonic, Country, etc.)
     - Format (Format covers any current media storage methodology in place. LP, CD, Cassette, mp3, etc)
     - Label (This is the label that produced the recording)
     - Label_Number (This is the inventory value that the Label assigned the recording)
     - Recording_Mode (This is the type of recording. One of Stereo, Mono or Unknown. Typically a search on AI such as chatGPT or Google can determine the mode based on the label_number, year of release and either Title or Artist)
     - DBX_Encode (this is reserved for recordings that may have been reissued and encoded with the DBX signal processing/compression methodology. Users need a DBX decode/encode box to play recordings, typically LP.)

Another field that is NOT extracted by the python script is "Reissue". Although one can filter for this in the discogs key-word search, extracting the info from the discogs database is unreliable at this time. The same is true for the "DBX" field. Sometimes the python script will be able to extract this (information), but if the recording information does NOT list "DBX encoding" in the recording information (there may be multiple entries that are listed besides the Master release).